+++ src/test/java/roomescape/payment/fixture/PaymentFixture.java
@@ -4,6 +4,7 @@ import static roomescape.reservation.fixture.ReservationFixture.SAVED_RESERVATIO
 import static roomescape.time.fixture.DateTimeFixture.TOMORROW;
 import static roomescape.time.fixture.ReservationTimeFixture.RESERVATION_TIME_10_00_ID_1;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import roomescape.payment.domain.Payment;
 
@@ -11,5 +12,5 @@ public class PaymentFixture {
 
     public static final Payment PAYMENT = new Payment(
             1L, SAVED_RESERVATION_1.getId(), "paymentKey", "orderId", "orderName",
-            1000L, LocalDateTime.of(TOMORROW, RESERVATION_TIME_10_00_ID_1.getStartAt()));
+            BigDecimal.valueOf(1000), LocalDateTime.of(TOMORROW, RESERVATION_TIME_10_00_ID_1.getStartAt()));
 }
